Refactor: Fix "privileges" misspelling
