+++ core-windows/src/router.cpp
@@ -33,7 +33,7 @@
 #include "../lib/json/json.hpp"
 #include "auth/authbasic.h"
 #include "ping/ping.h"
-#include "cloud/previleges.h"
+#include "cloud/privileges.h"
 
 
 using namespace std;
@@ -95,8 +95,8 @@ namespace routes {
 
                     bool permission = true;
 
-                    if(previleges::getMode() == "cloud") {
-                        if(!previleges::checkPermission(modfunc)) permission = false;
+                    if(privileges::getMode() == "cloud") {
+                        if(!privileges::checkPermission(modfunc)) permission = false;
                     }
 
                     if(permission) {
